Initial Upload
Keep this code in a safe place, it works well.
It changes the state of the GP pins based on the value of a PWM signal
from an Rx
It was for a video switcher but in fact when I came to design the
Project board layout it was too big for my purpose so I will order an
off the shelf video switcher.